### 2018-06-10 3acf635
streamline manifest updating, should close #3
