### 2018-06-11 5c92c87
fixes for editor checks, thanks @annakrystalli
