+++ man/pb_delete.Rd
@@ -7,7 +7,7 @@
 pb_delete(repo = guess_repo(), tag = "latest", file, .token = get_token())
 }
 \arguments{
-\item{repo}{Repository name in format "owner/repo". Will gess the current
+\item{repo}{Repository name in format "owner/repo". Will guess the current
 repo if not specified.}
 
 \item{tag}{tag for the GitHub release to which this data should be attached.}
