### 2018-06-11 #220 @cboettig
@annakrystalli Thanks!

- [x] I just learned about `devtools::spell_check()`, very cool.  spelling errors fixed .
- [x] review badge added
- [x] I added a mention of `usethis::edit_r_environ()` to README
- [ ] That test should be getting skipped if you're running locally, because you don't have my GitHub token...  I'm guessing you have your own `GITHUB_TOKEN` or `GITHUB_PAT` already set?  Can you try running checks without a `.Renviron` file that sets either env var?

I'm not sure how to change the `skip` command to handle that case; currently it just skips the authentication-requiring tests if the token is undefined (`""`) (i.e. so it will skip on CRAN; travis runs these tests since I can give travis an encrypted token.  Note the coverage will be lower than what [codecov](https://codecov.io/github/cboettig/piggyback?branch=master) shows if you test locally due to these skips.
