### 2018-06-12 #220 @annakrystalli
Aha 🤔, not sure how feasible it seeks but it feels like the only people that test should work for are the repo owner and any collaborators. So perhaps the test condition check could include a check for the gh user to match one of `owner` or `collaborator`?

Also, all tests passing after disabling `GITHUB_PAT` 👌
