### 2018-06-12 #220 @annakrystalli
Also, current coverage,
```
piggyback Coverage: 0.36%
R/push-pull.R: 0.00%
R/utils.R: 0.00%
R/gh.R: 0.77%
```
As you mentioned, some tests didn't run. What sort of coverage are you getting?
