### 2018-06-12 #220 @annakrystalli
Sorry just found link to codecov. Ignore comment above.
