### 2018-06-12 #220 @richfitz
> I'm not sure how to change the skip command to handle that case; currently it just skips the authentication-requiring tests if the token is undefined ("") (i.e. so it will skip on CRAN; travis runs these tests since I can give travis an encrypted token

@cboettig - I use "opt in" environment variables for tests like this, something like `<PACKAGENAME>_USE_<FEATURE>` and set that in my .Renviron and .travis.yml, then in the helper functions in test that have an equivalent `skip_if_not_<feature>` function that skips the test if the variable is not set to true
