### 2018-06-12 #220 @mpadge
Or universal local ones: I just set `MPADGE_LOCAL=true`. If I've got that, then `GITHUB_PAT` must be mine.
