### 2018-06-12 3140b60
use a second env var to indicate my PAT
this clever work-around is from @mdpage :shippit:
Now to let's see if I haven't bonked my travis or appveyor configs...
