+++ appveyor.yml
@@ -14,6 +14,7 @@ cache:
   - C:\RLibrary
 
 environment:
+  CBOETTIG_TOKEN="TRUE"
   my_variable:
     secure: OilSsZNqHHS5tRmeTZTVgCW7f7wqWhsMu5pDDj0J0UKfs9za+CyJH2Gqj9yDL1wQ
 
