+++ .travis.yml
@@ -6,3 +6,4 @@ after_success:
 env:
   matrix:
     secure: mJzI7ky/OpX4swsbWsX/Y7DG2ST5LXvbJ2E/IqGdi1QNt8c/IG20945ikWqotrOMRky4cc/CbWPbzGeTZgAFVas1IJ/K68GSWwzZjKzKp6jDy3YGCwRIREAaOp7BvAChq2IYe6KajAp8zgikIydq3beN3Saf9O4yPBXR3TZ/utowwXG9LvnHW8IYQeZDoE1dMf3rnfrfiOxXqsHO7yrLlIm0/2DVmcCsr7NeGTo2lZF/YrlFmozuZaetytCCqbB9HDC6goz+g0gWOcyI9hbW2YSzFPVUdaTsquYpDZdVaCMOKjEa1uLbZB264E2tVGZw8jm6ZYzVnuDP3WGIn9S76MAoEYvlrVXHTTPY8H5CQBQseVh5UWqoFh0TU4wSclvUKNiSHVgOew9QIlXNRfQp3cfF2JOoDQeRGzRDxeHG+I4lCVLc56ITN1HPlZGIDh79YEnqBcEpOIzHT73LG7AvZuULRHt/Ma5KFA+/A7CRXurPoxAOlndPZwJAW+w20C/38KqyY/ipz5NulIunAd7y1fovo1oeQvzwWa12tt6jlilKUi57V7b61i8WZ1YQ4KJslxQvDGY1Z9Zf8+bswCqNXMOtqb+4UJRnZ6RCOEGTdLCconjh9Mb3pCmzWp6xks4qpw2ty7xEGnBwZTQSAI66mnWDUFEHb5GwCy46YmJsw/U=
+    CBOETTIG_TOKEN="TRUE"
