### 2018-06-12 #220 @cboettig
Thanks all, that's brilliant.  An after a few boneheaded attempts I managed to pass the secure+public env vars to travis correctly this time.
