### 2018-06-13 #220 @coatless
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

Estimated hours spent reviewing: 1.5hrs

---

#### Environment Variable Access

The environment variable for the Personal Access Token (PAT),  `GITHUB_TOKEN`, is not the common way the PAT is stored across development packages. Generally, the PAT in environment variables is given as `GITHUB_PAT`. Thus, an additional system environment check is warranted.

Note, that `gh` and `devtools` both prefer `GITHUB_PAT`.

https://github.com/r-lib/gh/blob/8677f7d287ef76c33f33831ea09a6df89b31f88e/R/gh_request.R#L143-L146

https://github.com/r-lib/devtools/blob/26c507b128fdaa1911503348fedcf20d2dd30a1d/R/github.R#L86-L106

I think it would also be advantageous to mimic the `github_pat` function from `devtools` in place of `get_token()`:

https://github.com/cboettig/piggyback/blob/5c92c8753eda0c50f721ee2b39d0f03c8bc24b80/R/gh.R#L281-L283

#### Package API

The API for the package dances back'n'forth between treating the _repository_ as the data and the data as a file. I can understand the use-cases for both;  but, would it be better practice to emphasize the repository as the data? This would simplify the interface to mimic pre-existing APIs, e.g. `devtools::install_*` and `download.files(url, dest)`. Alternatively, consider making the API more verbose, e.g. `pb_download_repo(repo)` or `pb_download_repo_file(repo, file)`

Package API:
``` r
  library("piggyback")
  lsf.str("package:piggyback")
#> pb_delete : function (repo = guess_repo(), tag = "latest", file, .token = get_token())
#> pb_download : function (file = NULL, dest = ".", repo = guess_repo(), tag = "latest",
#>     overwrite = TRUE, ignore = "manifest.json")
#> pb_download_url : function (file = NULL, repo = guess_repo(), tag = "latest")
#> pb_list : function (repo = guess_repo(), tag = "latest", ignore = "manifest.json")
#> pb_new_release : function (repo = guess_repo(), tag, commit = "master", name = tag,
#>     body = "Data release", draft = FALSE, prerelease = FALSE, .token = get_token())
#> pb_pull : function (repo = guess_repo(), tag = "latest", overwrite = TRUE, manifest = ".manifest.json")
#> pb_push : function (repo = guess_repo(), tag = "latest", overwrite = TRUE, manifest = ".manifest.json")
#> pb_track : function (glob)
#> pb_upload : function (file, repo = guess_repo(), tag = "latest", name = NULL,
#>     overwrite = FALSE, .token = get_token())
```

#### Function Parameters

Another notable issue is the placement of the positional parameter _after_ defaults.

e.g. `file` in `pb_delete`
```r
pb_delete(repo = guess_repo(), tag = "latest", file, .token = get_token())
```

Consider putting `file` first or indicating missingness with `file = NA`.

#### Function Exports

It may be beneficial to also export and document `guess_repo()` as that is prominently featured in function signatures.

https://github.com/cboettig/piggyback/blob/5c92c8753eda0c50f721ee2b39d0f03c8bc24b80/R/push-pull.R#L261-L268

Outside of that, it seems like there are a few artifact of an older API in the documentation examples and some exported functions are missing documentation:

Missing ex:

- https://github.com/cboettig/piggyback/blob/5c92c8753eda0c50f721ee2b39d0f03c8bc24b80/R/gh.R#L21

- https://github.com/cboettig/piggyback/blob/5c92c8753eda0c50f721ee2b39d0f03c8bc24b80/R/gh.R#L76
- https://github.com/cboettig/piggyback/blob/5c92c8753eda0c50f721ee2b39d0f03c8bc24b80/R/gh.R#L234

Outdated:
- https://github.com/cboettig/piggyback/blob/5c92c8753eda0c50f721ee2b39d0f03c8bc24b80/R/gh.R#L132
