### 2018-06-14 #220 @annakrystalli
Wow! I'm sure this might be some sort of record @coatless! 👏 Thanks so much for such a super speedy response!

@cboettig  I'm still looking for the second reviewer. Feel to wait and respond to both reviews together or respond to @coatless anytime before.
