### 2018-06-15 ac5acf0
fix documentation examples as @coatless points out in review
