### 2018-06-15 #220 @cboettig
@coatless thanks, this is great!  How'd your use case go?

I've fixed the missing and outdated examples.  I've made most of the examples into `donttest` and made sure they run successfully with `devtools::check(cran = FALSE, run_dont_test = TRUE)`, though that requires my PAT since the tests upload data to `cboettig/piggyback`.  Not sure if there's a better strategy for documentation than that -- obviously this means copy-pasting examples won't work, but examples need a repo they can upload to.  Ideas / suggestions welcome.

Re the "file" vs "repo" interface, that's very intentional -- but perhaps I need to document it better.  If I'm just trying to debug an example on some data, I'd rather use a file-based interface to quickly share the file over, say,  a generic "scratch" repository (regardless of my current working repo, or if I'm not "in" any repo).   Locally I can do:

```r
pb_upload("mtcars.tsv.gz", repo  = "cboettig/scratch")
pb_download_url("mtcars.tsv.gz", repo  = "cboettig/scratch")
```

Then  I can post a question like: "Hey Jim, why does this fail:"

```r
read.table("https://github.com/cboettig/scratch/releases/download/data/mtcars.tsv.gz")
```

This needs the more file-based workflow of upload/download/download_url/list/delete.  Other times, I'm working in a particular repository with data connected to that project, and I really want the more "git-lfs" style workflow with the ability to track, say, all `*.tif` or `*.tsv.gz` files.    Does that make sense?  Can I do a better job of explaining that somewhere?

Re `GITHUB_PAT` -- as you probably saw, `piggyback` will recognize that name too.  But I do think that choice was one of Hadley's rare mistakes, `GITHUB_TOKEN` appears in some 86K code files on GitHub vs only 2K that use `GITHUB_PAT`, and  `gh` package also recognizes both.  I probably should do something like devtools does to provide a default public token though.  I would import the `devtools` method directly but I'm hesitant to do so while the `devtools` diaspora still seems underway; not sure where that routine will end up.  I've tried instead to build on `usethis` routines here whenever possible.  Perhaps @jennybc has insight here?

Re exporting `guess_repo()`, I agree that's a useful pattern, I wish I could import it from `usethis` or `git2r` myself!, but I think I should keep it internal precisely because it's quite general -- I really wouldn't want people introducing a revdep on `piggyback` just for that function, and I don't want to guarantee that method is robust and general, rather than an internal method whose behavior might change.
