### 2018-06-19 43992c7
patch tracking of directory paths
