### 2018-06-20 1a91657
avoid pb_pull error when dir doesn't exist.
create_manifest should no longer error when inspecting non-existant local directory listed in tracked files.

closes #6
