### 2018-06-21 fcb204b
patch pb_list to work with no ignored files
handle the case of i = integer(0)
closes #7
