### 2018-06-21 985f1e7
somewhat better parsing of .pbattributes
