+++ NAMESPACE
@@ -14,7 +14,9 @@ importFrom(crayon,blue)
 importFrom(crayon,green)
 importFrom(fs,dir_create)
 importFrom(fs,dir_ls)
+importFrom(fs,path_filter)
 importFrom(fs,path_join)
+importFrom(fs,path_rel)
 importFrom(gh,gh)
 importFrom(git2r,discover_repository)
 importFrom(git2r,remote_url)
@@ -29,5 +31,6 @@ importFrom(httr,upload_file)
 importFrom(httr,write_disk)
 importFrom(jsonlite,toJSON)
 importFrom(jsonlite,write_json)
+importFrom(magrittr,"%>%")
 importFrom(usethis,proj_get)
 importFrom(usethis,use_git_ignore)
