### 2018-06-21 #220 @cboettig
@annakrystalli Karthik says badge server got restarted for upgrades, looks like it is back now.
