### 2018-06-26 #220 @mpadge
Package does not seem to have contributing guidelines in README or as separate file

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

There are no references. Note that I previously submitted a reference-less `paper.md` to JOSS and was asked to please provide at least one reference.

---

This is a great and (for me at least) highly necessary package that makes my life so much easier. I am currently collaborating on a large project with multiple repos hosting code to generate very large data sets for which this package has provided an ideal solution to prevent repo-bloat. While this could all be done manually and laboriously, it is now possible to simply send links to collaborators who've got better things to do that learn the intricacies of git(hub), and write in the `README`:
library (our_data_packge)
download_all_data ()
```
That's awesome! Thanks @cboettig!

#### My major criticism

To get the negative out of the way first: The package as it stands is in many ways likely to be barely usable by non-coders because it issues almost no informative error messages. Almost all arguments appear to be passed un-checked directly to the github API which then returns a variety of largely unexpected and unintelligible `http` request errors.

Providing useful error message is crucial for productive use of any package, and would in this case have make the job of reviewing it considerably easier. Let me illustrate what I consider to be some inadequately processed errors:

```
pb_new_release(repo = "mpadge/myrepo", tag = "v1", .token = NULL)
#   Not Found (HTTP 404).
```
Please issue a useful message about the necessity of local tokens.
```
pb_upload ("non-existent-file.Rds", repo = "mpadge/myrepo")
# Error in httr::upload_file(file) : file.exists(path) is not TRUE
```
Please intercept such errors prior to actually calling `httr` and relying on that package to issue your errors.
```
pb_upload ("existent-file.Rds", repo = "mpadge/myrepo",
           tag = "non-existent-tag")
#                 Not Found
```
That last one is particularly unhelpful, and would be much more useful with perhaps a dialog starting
# Tag 'v2' does not exist in that repository. Would you like it to be created?
```
(With suitable control over, or default values for, draft and pre-release options.)
pb_upload (file = "this-file-already-uploaded.Rds")
#   Unprocessable Entity (WebDAV; RFC 4918) (HTTP 422).
pb_upload (file = "this-file-already-uploaded.Rds", overwrite = TRUE) # okay
```
The first of these should generate a message/warning about the `overwrite` parameter, rather than a 422 error. Finally,
```
pb_new_release (tag = "this-tag-already-exists")
#   Unprocessable Entity (WebDAV; RFC 4918) (HTTP 422).
```
In short: Please pre-empt and process all errors within this package, rather than relying on external packages to detect and issue your error messages.

#### Back to the positive

Okay, now that that's out of the way, let me return to the positive tone in which I began. This really is a great and hugely useful package, and I hope that the remainder of this review can somehow contribute to making it even better.

Before proceeding, I note a general issue raised by @coatless regarding (in his words), the dance "between the *repository* as the data and the data as a file." I would extend that dance to a parallel and unavoidable one between git and github, and specifically between tags and releases. The package utilises the mechanism of github releases to aid enable clean interfaces with locally-hosted git repositories which only understand tags. I am unsure whether this is a package designed to interface with a github repo (via releases) or a local repo (via tags), although do not the useful section of the README/vignette in this context. Consider nevertheless that `pb_new_release()` will construct both a new tag on github, and a corresponding new release, yet this tag is not fetched locally, and so my local version of the repo will not know this tag.

While the package clearly has a primary github (rather than git) focus, it is nevertheless largely built around the `guess_repo()` function - which in my opinion, to strongly concur with @coatless, really ought to be exported - and this function is actually built on entirely local (git) queries (via `git2r::discover_repository()`. Given this, I suggest it would be useful to at least have some kind of optional (if not default) trigger to fetch all remote tags created with `pb_new_release()` to the local instance.

@cboettig also has a comment in the current code of the `gh.R/pb_list()` function:
> Should we report the tag name too? x$tag_name?

To which I absolutely suggest Yes. It is perhaps useful to describe my current project to indicate perhaps a kind of use case not directly envisioned here. We have a variety of data representing a variety of stages from raw data to final output. There are several processing stages, and each of these now has - thanks to the ease and brilliance of this package - its own tag and corresponding github release. We therefore generally need to know at all times what the tags actually are, and with which tags each file is associated. This currently requires a command-line git interface to discern the tags, and the `piggyback` interface to list the files. Having these two combined within a single `pb_list()` call would be supremely useful here.

I also feel that `pb_list()` should by default list all files, enabling the kind of insight I just mentioned, with `tag` then restricting the view to only those associated with that tag (whether `latest` or not). It would also be very useful to have a simple `pb_tags()` function that listed all tags on the github repo.

#### Primary functionality

The package is divided between two modes, the explicit `pb_up/download()` and `pb_push/pull()`. I find both of these very useful, yet suspect that a merging of the two may ultimately be even more useful. The main difference actually seems to be the creation by the latter of a local `.pbattributes` file which is then used to track all files that are or can be pushed and pulled, enabling wholesale calls to `pb_push/pull()` without having to explicitly specify file names. On the other hand, `pb_up/download()` always require explicit filenames, and these calls never interact with `.pbattributes` at all.

Again, I do not think my current use case is likely to be particularly unusual, but we simply have several files of which only some are to be `push/pull`-ed or `up/download`-ed. This currently requires the latter, explicit interfaces, yet relying on these then prevents a simple `pb_pull()` call to download all data, which is a shame. One easy solution would be for `pb_up/download()` to also add entries to `pb_push/pull()`, so that `pb_pull()` could be used, but I suspect a nicer solution might ultimately be to have a single pair of calls (say, `pb_up/download()`, although i have no preference for nomenclature either way), such that `pb_upload ("*.dat")` functions as `pb_push()` currently does; `pb_upload("x.dat")` uploads that file only **and** adds the name to `.pbattributes`, and `pb_upload()` uploads all files listed on (or matching globs in) `.pbattributes`. Same for downloads.

#### Miscellaneous other notes

#### pb_delete

The following call simply and mysteriously completes silently:
pb_delete ("these-data-do-not-exist.Rds")
```
As does this one:
pb_delete ("these-data-do-exist.Rds")
```
Please issue an message in the first instance. It would also be useful to be able to:
pb_delete (tag = "v1")
```
with the expected results then manifest both locally and on github.

#### pb_push/pull

Note that `pb_push/pull()` generally behave as expected, yet if a file is deleted locally, then calling `pb_pull()` no longer displays the name of the file prior to displaying the progress bar.  This inconsistency ought to be rectified.

Further note that if that file is again removed locally, then calling `pb_push()` simply completes silently. It would be useful in this instance to check any files explicitly listed in `.pbattributes`, and to issue some kind of informative messages that there are no local files to push.

I also suspect that some more thought might be productively put into exactly what role the `.pbattributes` file plays. If files are deleted locally, should they also be removed from `.pbattributes`? In this case, a call to `pb_delete()` could easily do all of these.

#### GITHUB_TOKEN

See discussion of nomenclature [here](https://github.com/ropensci/ghql/issues/11), and my personally opinionated view that "GITHUB_PAT" is obfuscatory - "TOKEN" is a word; "PAT" is not.

I note simply that when this package goes to CRAN, all tests will be skipped. I'm not sure of their current methods to actually examine the structure of tests, but it is certainly no longer acceptable to have a package without tests. If they actually examine whether tests are run (rather than just written), then the tests may need to be modified (purely in regard to CRAN, not at all in regard to this review).

---

Thanks @cboettig for a supremely useful and (lack of error messages notwithstanding) easy-to-use package. I look forward to it only become even easier to use through this review process.
