### 2018-07-04 ropensci/piggyback#8 @mpadge
Warn when downloading/pulling older versions of files
With respectful acknowledgment of your mistrust of timestamps, one place where these would actually be useful would be in order to implement a warning when downloading/pulling older versions of files, and consequently overwriting newer versions that may have since been created locally. That would be a really useful safety mechanism (for my current workflow at least). We're throwing files between lots of different people doing different tasks, and inadvertent overwrites do happen, and are potentially painful.
