### 2018-07-04 ropensci/piggyback#8 @cboettig
Yeah, overwrites are definitely and issue so that's a good idea.

Maybe we should go one further and simply provide the option of `use_timestamps` = `FALSE`/`warn`,/`TRUE` to `push`/`pull`.  `warn` would do as you say, `FALSE` is the current behavior, and `TRUE` would only push/pull those files with differing hashes AND older timestamps.

Perhaps my mistrust of timestamps is not so warranted -- mostly arises from the fact that switching branches changes the timestamp, and from work on poorly maintained servers with drifting clocks long ago.  I guess since these files are not git managed, they would not be impacted by the latter.  A more empirically grounded opinion on timestamps might be called for.  (I agree that managing the hash tree a la git is the 'real' solution to version tracking, but I'm not going to implement that by hand).
