### 2018-07-04 #220 @cboettig
@mpadge Thanks for this, fantastic review.

- Completely agree about the error message issues (and associated tests), will fix that once I resolve the interface issues first.

- Re TOKEN vs PAT, agree on TOKEN (as I had stated in thread you linked, that's the dominant standard elsewhere), but lots of R users adopted PAT thanks to devtools, `piggyback` supports both (as does `usethis` and `gh`), so hopefully we can call that a truce.

- tests: yeah, I'm supremely mistrustful of CRAN's servers to test things that require network connections, though I could add some more trivial tests.  In principle all the `tests-without-auth.R` should work on CRAN, so maybe I should just let CRAN attempt those, but you know how it is.

- interface.  Yeah, this is the big issue I should address first, since it will probably impact how all the others are done.  I think I could use some more input on this:

  - I agree that `push`/`pull` are essentially `upload`/`download` with a file list generated from `.pbattributes`, even though they aren't quite written like that currently (currently only `push`/`pull` handle the manifest / hashes, as you know).

  - I actually still cannot decide if the `download` / `upload` interface could support the `manifest.json` bit (at least as it is currently implemented).  On the face of it, this sounds obvious and trivial to check hashes before downloading, but keeping an accurate manifest on GitHub isn't trivial, particularly if we're managing file-by-file instead of as a suite of files -- every `download` command would have to first download the manifest, and every upload would have to download, merge, and upload a merged manifest.  `delete` would also have to download and upload a `manifest`, etc.  (Of course none of this would be necessary if the GitHub API would compute hashes for us).  I just worry this increases the error surface of what are otherwise super simple low-level functions for often very little gain (maybe I'm spoiled by high bandwidth; I've been impressed that GitHub is willing to download assets at > 200 MB/s, Zenodo doesn't get near those rates).  More thoughts on this would be appreciated.
