### 2018-07-04 ropensci/piggyback#8 @mpadge
That sounds like a great solution. And yeah, the whole idea of `piggyback` is to avoid clogging up repos with the piggybacked files, so one can surely assume they very generally won't be part of a git tree, so should not have time stamps altered by any git operations. This would - for me - be a great boost!
