### 2018-07-07 9788cdc
use links in package level documentation
