+++ R/gh.R
@@ -11,10 +11,13 @@
 #' Should be a directory or a list of filenames the same length as `file` vector.
 #' Can include paths to files, but any directories in that path must already exist.
 #' @param tag tag for the GitHub release to which this data is attached
-#' @param overwrite default `TRUE`, should any local files of the same name be overwritten?
+#' @param overwrite Should any local files of the same name be overwritten? default `TRUE`.
 #' @param ignore a list of files to ignore (if downloading "all" because `file=NULL`).
 #' by default, "manifest.json" is ignored as this file is created and used only for
-#' avoiding redundant file transfer by `pb_push` and `pb_pull`
+#' avoiding redundant file transfer by [pb_push()] and [pb_pull()]
+#' @param use_timestamps If `TRUE`, then files will only be downloaded
+#' if timestamp on GitHub is newer than the local timestamp (if `overwrite=TRUE`).
+#' Defaults to `TRUE`.
 #' @importFrom httr GET add_headers write_disk
 #' @importFrom gh gh
 #' @importFrom fs dir_create
@@ -31,39 +34,64 @@ pb_download <- function(file = NULL,
                         repo = guess_repo(),
                         tag = "latest",
                         overwrite = TRUE,
-                        ignore = "manifest.json"){
+                        ignore = "manifest.json",
+                        use_timestamps = TRUE){
 
-  ## FIXME revisit using path escape logic!!
+
 
   x <- release_info(repo, tag)
-  id <- vapply(x$assets, `[[`, integer(1), "id")
-  gh_file_names <- vapply(x$assets, `[[`, character(1), "name")
-  file_names <- local_filename(gh_file_names)
+
+  # drop ignore
+  # enforce timestamp
+  # handle file = NULL
+  # return ids
+
+  df <- data.frame(
+    id = vapply(x$assets, `[[`, integer(1), "id"),
+    file_name = local_filename(vapply(x$assets, `[[`,
+                                      character(1), "name")),
+    timestamp = as.POSIXct(vapply(x$assets, `[[`,
+                                  character(1), "updated_at")),
+    stringsAsFactors = FALSE)
 
   if(!is.null(file)){
-    i <- which(file_names %in% file)
-    id <- id[i]
+    i <- which(df$file_name %in% file)
+    if(length(i) < 1)
+      stop(paste("file(s)",
+                 paste(crayon::blue(file), collapse=" "),
+                       "not found in repo",
+                       crayon::blue(repo)))
+
+    df <- df[i,]
   } else {
-    i <- which(file_names %in% ignore)
-    if(length(i) > 1){
-      file_names <- file_names[-i]
-      id <- id[-i]
+    i <- which(df$file_name %in% ignore)
+    if(length(i) >= 1){
+      df <- df[-i,]
     }
-    file <- file_names
+    file <- df$file_name
   }
+
+
+
   ## if dest not provided, we will write
-  if(length(dest) <= 1){
-    i <- which(file_names %in% file)
+  if(length(dest) == 1){
+    i <- which(df$file_name %in% file)
     ## Make sure dest dir exists!
-    dest <- fs::path_rel(file.path(dest, file_names[i]))
+    dest <- fs::path_rel(file.path(dest, df$file_name[i]))
     fs::dir_create(fs::path_dir(dest))
-
   }
 
-  for(i in seq_along(id)){
+  if(use_timestamps){
+    local_timestamp <- fs::file_info(dest)$modification_time
+    update <- df$timestamp > local_timestamp
+    update[is.na(update)] <- TRUE # we'll download if missing locally
+    df <- df[update,]
+  }
+
+  for(i in seq_along(df$id)){
[…]
