+++ man/pb_download.Rd
@@ -5,7 +5,8 @@
 \title{Download data from an existing release}
 \usage{
 pb_download(file = NULL, dest = ".", repo = guess_repo(),
-  tag = "latest", overwrite = TRUE, ignore = "manifest.json")
+  tag = "latest", overwrite = TRUE, ignore = "manifest.json",
+  use_timestamps = FALSE)
 }
 \arguments{
 \item{file}{name or vector of names of files to be downloaded. If \code{NULL},
@@ -24,7 +25,10 @@ the current repo if not specified.}
 
 \item{ignore}{a list of files to ignore (if downloading "all" because \code{file=NULL}).
 by default, "manifest.json" is ignored as this file is created and used only for
-avoiding redundant file transfer by \code{pb_push} and \code{pb_pull}}
+avoiding redundant file transfer by \code{\link[=pb_push]{pb_push()}} and \code{\link[=pb_pull]{pb_pull()}}}
+
+\item{use_timestamps}{default \code{FALSE}. If \code{TRUE}, then files will only be downloaded
+if timestamp on GitHub is newer than the local timestamp.}
 }
 \description{
 Download data from an existing release
