### 2018-07-09 f941d00
tidy up behavior when nothing to download
