+++ tests/testthat/test-on-cran.R
@@ -1,0 +1,11 @@
+
+testthat::test_that(
+  "we can download a requested file from the latest release", {
+    pb_download( file = "data/mtcars.tsv.gz",
+                 repo = "cboettig/piggyback",
+                 dest = tempdir())
+
+    f <- fs::path(tempdir(), "data/mtcars.tsv.gz")
+    testthat::expect_true(file.exists(f))
+    unlink(f)
+  })
