### 2018-07-12 e01837a
use spelling::spell_check_setup
