+++ DESCRIPTION
@@ -32,6 +32,7 @@ Imports:
     clisymbols,
     magrittr
 Suggests:
+    spelling,
     readr,
     covr,
     testthat,
@@ -41,3 +42,4 @@ Suggests:
 VignetteBuilder: knitr
 RoxygenNote: 6.0.1.9000
 Roxygen: list(markdown = TRUE)
+Language: en-US
