### 2018-07-16 #220 @annakrystalli
Hello @cboettig and @mpadge 👋.

Just wondering if any progress had been made with the issues highlighted in the review and feedback requested by @cboettig?
