### 2018-07-20 b490116
only match files; i.e. objects we can md5sum
