### 2018-07-20 d92239a
Implement timestamps (#8)
using timestamps is now the default for upload/download, but not for push/pull.
