+++ man/pb_push.Rd
@@ -5,7 +5,7 @@
 \title{Push data to GitHub}
 \usage{
 pb_push(repo = guess_repo(), tag = "latest", overwrite = TRUE,
-  manifest = ".manifest.json")
+  manifest = ".manifest.json", use_timestamps = FALSE)
 }
 \arguments{
 \item{repo}{Name of the repo on GitHub (\code{owner/repo}, i.e.
@@ -21,6 +21,10 @@ not match? default \code{TRUE}.}
 \item{manifest}{name of the local manifest file. Note: A leading dot
 (i.e. indicating a hidden file) in the manifest name will be removed
 from the name used on the GitHub asset list.}
+
+\item{use_timestamps}{If \code{TRUE}, then files will only be uploaded/downloaded
+if timestamp of target is newer than the existing version.  Default is \code{FALSE},
+see details.}
 }
 \description{
 Push all currently tracked data files to GitHub.  Only files identical
@@ -32,6 +36,18 @@ GitHub versions when pushing.
 \details{
 Will only upload tracked files, as identified by the local
 manifest.  Add files to tracking with \code{\link{pb_track}} first.
+
+By default, use_timestamps is false in \code{\link[=pb_pull]{pb_pull()}} amd \code{\link[=pb_push]{pb_push()}} since
+these methods are designed to use the manifest, which relies on hashes
+rather than timestamps to decide if a file has changed.  You can use
+timestamps and hashes together, but note that timestamps may not be
+as reliable, particularly if your files are being moved or copied
+some other way without actually being updated.
+
+Hash comparisons can be more reliable, but unlike timestamps, have no
+directionality -- we cannot tell which file is most recent.  Checking
+hashes therefore only ensures we not bother uploading or downloading
+a file identical to one we already have.'
 }
 \examples{
 \dontrun{
