+++ man/pb_push.Rd
@@ -37,15 +37,15 @@ GitHub versions when pushing.
 Will only upload tracked files, as identified by the local
 manifest.  Add files to tracking with \code{\link{pb_track}} first.
 
-By default, use_timestamps is false in \code{\link[=pb_pull]{pb_pull()}} amd \code{\link[=pb_push]{pb_push()}} since
+By default, use_timestamps is false in \code{\link[=pb_pull]{pb_pull()}} and \code{\link[=pb_push]{pb_push()}} since
 these methods are designed to use the manifest, which relies on hashes
 rather than timestamps to decide if a file has changed.  You can use
 timestamps and hashes together, but note that timestamps may not be
 as reliable, particularly if your files are being moved or copied
 some other way without actually being updated.
 
-Hash comparisons can be more reliable, but unlike timestamps, have no
-directionality -- we cannot tell which file is most recent.  Checking
+Hash comparisons can be more reliable, but unlike timestamps, are not
+directional -- we cannot tell which file is most recent.  Checking
 hashes therefore only ensures we not bother uploading or downloading
 a file identical to one we already have.'
 }
