### 2018-07-24 00b28c2
improve testing of errors
