+++ paper.bib
@@ -8,10 +8,10 @@ authors = {{GitHub Inc.}}
 }
 
 
-@MISC{GitLFS,
+@MISC{API3,
 url = {https://developer.github.com/v3/},
 canonicaluri = {\url{https://developer.github.com/v3/}},
-title = {Git LFS},
+title = {GitHub API version 3},
 YEAR = {2018},
 howpublished = {\url{https://developer.github.com/v3/}},
 authors = {{GitHub Inc.}}
