+++ R/gh.R
@@ -16,6 +16,11 @@
 #' @param use_timestamps If `TRUE`, then files will only be downloaded
 #' if timestamp on GitHub is newer than the local timestamp (if `overwrite=TRUE`).
 #' Defaults to `TRUE`.
+#' @param show_progress logical, should we show progress bar for download? default TRUE.
+#' @param .token GitHub authentication token. Typically set from an environmental
+#' variable, e.g. in a `.Renviron` file or with `Sys.setenv(GITHUB_TOKEN = "xxxxx")`,
+#' which helps prevent
+#' accidental disclosure of a secret token when sharing scripts.
 #' @importFrom httr GET add_headers write_disk
 #' @importFrom gh gh
 #' @importFrom fs dir_create
@@ -33,11 +38,17 @@ pb_download <- function(file = NULL,
                         tag = "latest",
                         overwrite = TRUE,
                         ignore = "manifest.json",
-                        use_timestamps = TRUE){
+                        use_timestamps = TRUE,
+                        show_progress = TRUE,
+                        .token = get_token()){
+
+  progress <- httr::progress("down")
+  if(!show_progress){
+    progress <- NULL
+  }
 
 
-
-  x <- release_info(repo, tag)
+  x <- release_info(repo, tag, .token)
   df <- rectangle_info(x)
 
   if(!is.null(file)){
@@ -84,7 +95,8 @@ pb_download <- function(file = NULL,
                       x$repo,
                       id = df$id[i],
                       destfile = dest[i],
-                      overwrite = overwrite)
+                      overwrite = overwrite,
+                      progress = progress)
 
   )
 }
@@ -103,9 +115,11 @@ pb_download <- function(file = NULL,
 #' }
 pb_download_url <- function(file = NULL,
                             repo = guess_repo(),
-                            tag = "latest"){
+                            tag = "latest",
+                            .token = .token){
 
-  x <- release_info(repo, tag)
+
+  x <- release_info(repo, tag, .token)
 
 
   gh_file_names <- vapply(x$assets, `[[`, character(1), "name")
@@ -126,14 +140,21 @@ gh_download_asset <- function(owner,
                               id,
                               destfile,
                               overwrite=TRUE,
-                              .token = get_token()
+                              .token = get_token(),
+                              progress = httr::progress("down")
                               ){
+  if(fs::file_exists(destfile) && !overwrite){
+    warning(paste(destfile, "already exists, skipping download.",
+                  "Set overwrite = TRUE to overwrite files."))
+    return(NULL)
+  }
+
   resp <- httr::GET(paste0("https://api.github.com/repos/", owner,"/",
                     repo, "/", "releases/assets/", id,
                     "?access_token=", .token),
                     httr::add_headers(Accept = "application/octet-stream"),
                     httr::write_disk(destfile, overwrite = overwrite),
-                    httr::progress("down"))
+                    progress)
   ## handle error cases? resp not found
   httr::stop_for_status(resp)
   invisible(resp)
@@ -151,11 +172,13 @@ gh_download_asset <- function(owner,
 #' `file` (i.e. filename without directory)
 #' @param overwrite overwrite any existing file with the same name already
 #'  attached to the on release?
-#' @param use_timestamps If `TRUE`, then files will only be downloaded
+#' @param use_timestamps logical, if `TRUE`, then files will only be downloaded
[…]
