+++ NAMESPACE
@@ -31,6 +31,7 @@ importFrom(httr,upload_file)
 importFrom(httr,write_disk)
 importFrom(jsonlite,toJSON)
 importFrom(jsonlite,write_json)
+importFrom(lubridate,as_datetime)
 importFrom(magrittr,"%>%")
 importFrom(usethis,proj_get)
 importFrom(usethis,use_git_ignore)

[…]
