### 2018-07-24 11278a8
skip if not a valid token
