### 2018-07-24 226b1d3
pb_delete should message when file does not exist
