+++ tests/testthat/test-with-auth.R
@@ -44,10 +44,10 @@ testthat::test_that("We can push and pull data",{
   setwd("piggyback")
 
   fs::dir_create("data")
-  readr::write_tsv(mtcars, "mtcars.tsv.gz")
-  readr::write_tsv(mtcars, "data/mtcars.tsv.xz")
+  readr::write_tsv(datasets::mtcars, "data/mtcars.tsv.gz")
+  readr::write_tsv(datasets::iris, "data/iris.tsv.xz")
 
-  pb_track(c("*.tsv.gz", "data/*"))
+  pb_track("data/*")
   testthat::expect_true(pb_push(repo = "cboettig/piggyback"))
   testthat::expect_true(pb_pull(repo = "cboettig/piggyback"))
 
