### 2018-07-24 b023994
avoid sink
warn instead of error if file does not exist
