+++ tests/testthat/test-without-auth.R
@@ -48,7 +48,8 @@ testthat::test_that(
   "we can list files", {
     testthat::skip_on_cran()
     testthat::expect_output(cat(
-      pb_list(repo = "cboettig/piggyback")), "iris.tsv.gz")
+      pb_list(repo = "cboettig/piggyback",
+              tag = "v0.0.1")), "iris.tsv.gz")
   })
 
 
