### 2018-07-27 #220 @mpadge
Thanks @cboettig. I'll take it our for a test ride in the coming week and report back then
