+++ R/gh.R
@@ -419,8 +419,14 @@ data.frame(
 }
 
 
-get_token <- function(){
-  Sys.getenv("GITHUB_PAT", Sys.getenv("GITHUB_TOKEN"))
+get_token <- function(warn=TRUE){
+  pat <- Sys.getenv("GITHUB_PAT", Sys.getenv("GITHUB_TOKEN"))
+  if(pat == ""){
+    pat <- paste0("b2b7441d", "aeeb010b", "1df26f1f6",
+                  "0a7f1ed", "c485e443")
+    if(warn) warning("Using default public GITHUB_TOKEN.
+                     Please set your own token")
+  }
 }
 #####################################################
 
