+++ tests/testthat/test-timestamps.R
@@ -4,7 +4,7 @@ testthat::test_that(
   "upload obeys timestamp only when newer",
   {
 
-    testthat::skip_on_travis() # No idea...
+    ## testthat::skip_on_travis() # No idea...
 
     testthat::skip_if(piggyback:::get_token() == "")
     testthat::skip_if(piggyback:::get_token() == "b2b7441daeeb010b1df26f1f60a7f1edc485e443")
