### 2018-08-15 #220 @mpadge
Thanks @cboettig for the great improvements to the package, and sorry for the slight delay getting back to you. The following comments are mere suggestions that in the subjective opinions of this reviewer might just help to make it even more useful and/or easy to use.

## Error messages and tags

As indicated by @cboettig, most previously hard-to-decipher error messages have been addressed, and the package is as a result considerably more user friendly. There nevertheless remain a couple of unhelpful error messages prompted by passing inappropriate arguments through to package dependencies, primarily `httr`. One simple one of these is:

```{r}
pb_upload ("non-existent-file.Rds", repo = "mpadge/myrepo")
```
This still passes non-existent file name through to `httr` and returns `file.exists(path) is not TRUE`. Please insert a simple catch at start of `pb_upload()`:
```{r}
if (!file.exists (file))
    stop ("file ", file, " does not exist")
```

The remainder may be more effectively addressed through first extracting all tags for a repository and then using those throughout to check each `tag` parameter. The extraction of tags is very simple using the graphql interface. My suggestion would be to ask @maelle if you might kindly borrow her `create_client` function from the [`ghrecipes` package](https://github.com/ropenscilabs/ghrecipes/blob/master/R/zzz.R):

```{r}
create_client <- function(){
  token <- Sys.getenv("GITHUB_GRAPHQL_TOKEN")
  if (token == "")
  {
    stop ("Please set your GITHUB_GRAPHQL_TOKEN environment variable.")
  } else
  {
    if (!exists ("ghql_gh_cli")) {
      ghql_gh_cli <- ghql::GraphqlClient$new (
        url = "https://api.github.com/graphql",
        headers = httr::add_headers (Authorization =
                                     paste0 ("Bearer ", token))
      )
      ghql_gh_cli$load_schema()
      ghql_gh_cli <<- ghql_gh_cli
    }
    return (ghql_gh_cli)
  }
}
```
A `get_tags` function could then simply be written like this:
```{r}
get_tags <- function (owner, repo)
{
    qtxt <- paste0 ('{
                    repository(owner:"', owner, '", name:"', repo, '") {
                        tags:refs(refPrefix:"refs/tags/", last:50) {
                            edges {
                                tag:node {
                                    name
                                }
                            }
                        }
                    }
    }')
    qry <- ghql::Query$new()
    qry$query('gettags', qtxt)
    dat <- create_client()$exec(qry$queries$gettags) %>%
        jsonlite::fromJSON ()
    if (is.null(dat$data$repository)) stop(dat$errors$message)

    dat$data$repository$tags$edges
}
get_tags ("mpadge", "junk")
```
This obviously introduces another dependency (`ghql`), but I suspect this package will likely be improved at some stage through incorporating `ghql` calls anyway. These tags can be cached somehow on the first call for a give repo, and then re-used throughout a session. A simple
```{r}
tags <- memoise::memoise (get_tags (owner, repo))
```
would do the trick (sorry, requiring yet another dependency, but I also really think the package will at some stage require some more sophisticated caching like this).

This would then allow the kind of prompt I suggested in the initial review along the lines of
```{r}
pb_upload (..., tag = "non-existent-tag")
# Tag 'non-existent-tag' does not exist in that repository. Would you like it to be created?
```
And it would happen instantly, without recourse to any API calls.

This would also enable error messages for the following call:
```{r}
pb_new_release(repo = "mpadge/junk", tag = "this-tag-already-exists")
```
At the moment this still returns the entirely illegible:
```{r}
# Error in pb_new_release(repo = "mpadge/junk", tag = "this-tag-already-exists") :
#   Unprocessable Entity (WebDAV; RFC 4918) (HTTP 422).
```

## other stuff

### pb_list

This function still does not return tag information - did you intend to incorporate that?  I would reiterate most of the comments I previously made regarding this function.

### pb_delete

This call now returns an appropriate error message:
```{r}
pb_delete (file = "non-existent-data", repo = "mpadge/junk", tag = "v1")
# non-existent-data not found on GitHub
```
As previously stated, it would still be nice to be able to delete an entire release. At the moment this is not possible:
```{r}
pb_delete (repo = "mpadge/junk", tag = "v1")
# Error in asset_filename(file) :
#   argument "file" is missing, with no default
```
Any comments on potentially enabling that?

### pb_track

This does indeed work as @cboettig suggests, but I note that `pb_upload` is not vectorised:

```{r}
pb_track ("data/*") # where * matches >= 2 files
pb_track () %>% pb_upload (repo = "mpadge/junk", tag = "v1")
# Error in vapply(elements, encode, character(1)) :
#   values must be length 1,
#  but FUN(X[[1]]) result is length 2
```
It would be very useful to simply make current `pb_upload` some kind of `pb_upload_one` function and then have `pb_upload` be a simple `vapply` call to that.

I'll leave my comments there for the moment, and await feedback from @cboettig
