### 2018-08-15 #220 @cboettig
Thanks @mpadge, really appreciate the detailed feedback and agree these things should be done.  My apologies that most of the things in "other stuff" look like oversights on my part, and should be straight forward to implement.

The only thing I'm on the fence on is the tags etc, not because I disagree about doing it, but that I think it's going to require more thought on the right way to do this.  Like you say, ultimately this package should really be caching some relatively detailed metadata about the package.  Of course this will require some more invasive surgery to implement...

I'm not sure `memoise` is the way to do this, since `memoise` invalidates the cache only when the function arguments change, and here the repo information is likely to change without that happening.  Still, it would be a convenient way to do so for sure, effectively making the cache persist through a session, so it might be an decent starting point.

I think I can get all the tags easily enough from the standard REST API, instead of `ghql`. Will have to dig into this...

Mark, I'd also like to ask your permission to list you as a contributor to the package, if that's okay with you?  Your detailed input has had a substantial contribution  to the package.
