### 2018-08-17 #220 @mpadge
:+1: Re: "ctb": :+1: of course! Thanks!

Re: tags + caching: Yeah, some invasive surgery will be required, and yeah, `memoise` is perhaps not ideal for the reasons you state, but then again `pb_new_release` is the only function that should change tags (at least until, or if, `pb_delete(tag = ...)` is incorporated), so consistent cache updating would just require re-running `memoise(get_tags())` within `pb_new_release`.

And yeah, API3 would also be entirely sufficient, because tag info will likely never run into rate restriction limits. I just got personally converted to API4 through @maelle's graphql evangelising, and don't plan on going back.

I hope all this back-and-forthing is okay with you, and I hope this doesn't come across as me being too picky. I really do just want to help make an already supremely useful package even cleaner and easier to use, and I strongly suspect that at least finding an initial system for interim caching will help make future package development much easier and smoother.
