### 2018-08-31 c16e5cc
vectorize pb_upload
