### 2018-08-31 11a6b58
Warn and exit call if file does not exist.
@mpadge I'm using a warning here instead of `stop` since the latter might be annoying in a vectorized list.  But happy to go to a `stop` if you really think that's the better behaviour.
