+++ man/pb_list.Rd
@@ -13,13 +13,14 @@ the current repo if not specified.}
 
 \item{tag}{tag for the GitHub release to which this data is attached}
 
-\item{ignore}{a list of files to ignore (if downloading "all" because \code{file=NULL}).
-by default, "manifest.json" is ignored as this file is created and used only for
-avoiding redundant file transfer by \code{\link[=pb_push]{pb_push()}} and \code{\link[=pb_pull]{pb_pull()}}}
+\item{ignore}{a list of files to ignore (if downloading "all" because
+\code{file=NULL}). By default, "manifest.json" is ignored as this file is
+created and used only for avoiding redundant file transfer by \code{\link[=pb_push]{pb_push()}}
+and \code{\link[=pb_pull]{pb_pull()}}}
 
-\item{.token}{GitHub authentication token. Typically set from an environmental
-variable, e.g. in a \code{.Renviron} file or with \code{Sys.setenv(GITHUB_TOKEN = "xxxxx")},
-which helps prevent
+\item{.token}{GitHub authentication token. Typically set from an
+environmental variable, e.g. in a \code{.Renviron} file or with
+\code{Sys.setenv(GITHUB_TOKEN = "xxxxx")}, which helps prevent
 accidental disclosure of a secret token when sharing scripts.}
 }
 \value{
