+++ tests/testthat/test-timestamps.R
@@ -4,8 +4,6 @@ testthat::test_that(
   "upload obeys timestamp only when newer",
   {
 
-    ## testthat::skip_on_travis() # No idea...
-
     testthat::skip_if(piggyback:::get_token() == "")
     testthat::skip_if(piggyback:::get_token() == "b2b7441daeeb010b1df26f1f60a7f1edc485e443")
     testthat::skip_if_not(as.logical(Sys.getenv("CBOETTIG_TOKEN", FALSE)))
