### 2018-09-03 81b2ef3
refactoring pb_list
A generic pb_info now returns information for assets on all releases as tabular data.  Will replace previous rectangularize method.
