### 2018-09-04 c876eb7
a minimal refactored version
