+++ R/gh.R
@@ -52,8 +52,6 @@ pb_download <- function(file = NULL,
   }
 
 
-  #x <- release_info(repo, tag, .token)
-  #df <- rectangle_info(x)
   df <- pb_info(repo, tag, .token)
 
   if(!is.null(file)){
@@ -131,17 +129,16 @@ pb_download_url <- function(file = NULL,
                             .token = get_token()){
 
 
-  x <- release_info(repo, tag, .token)
-
-
-  gh_file_names <- vapply(x$assets, `[[`, character(1), "name")
-  file_names <- local_filename(gh_file_names)
-
-  if(!is.null(file)){
-    i <- which(file_names %in% file)
+  df <- pb_info(repo, tag, .token)
+  if(is.null(file)){
+    return(df$browser_download_url)
+  }else if(file %in% df$file_name){
+    return(df[file == df$file_name, "browser_download_url"])
+  } else {
+    stop(paste("file", file, "not found in release", tag, "for repo", repo))
   }
-  urls <- vapply(x$assets, `[[`, character(1), "browser_download_url")
-  urls[i]
+
+
 }
 
 
@@ -252,8 +249,6 @@ pb_upload_file <- function(file,
     name <- fs::path_rel(file, start = dir)
   }
 
-  #x <- release_info(repo, tag, .token)
-  #df <- rectangle_info(x)
   df <- pb_info(repo, tag, .token)
 
 
@@ -315,7 +310,10 @@ local_filename <- function(x){
 
 #' List all assets attached to a release
 #' @inheritParams pb_download
-#' @return a character vector of release asset names, (normalized to local paths)
+#' @param tag which release tag do we want information for? If `NULL` (default),
+#' will return a table for all available release tags.
+#' @return a data.frame of release asset names, (normalized to local paths), release tag,
+#' timestamp, owner, and repo.
 #' @details To preserve path information, local path delimiters are converted to `.2f`
 #' when files are uploaded as assets.  Listing will display the local filename,
 #' with asset names converting the `.2f` escape code back to the system delimiter.
@@ -325,19 +323,11 @@ local_filename <- function(x){
 #' }
 #' @export
 pb_list <- function(repo = guess_repo(),
-                    tag = "latest",
+                    tag = NULL,
                     ignore = "manifest.json",
                     .token = get_token()){
-  x <- release_info(repo, tag, .token)
-  file_names <- vapply(x$assets, `[[`, character(1), "name")
-
-  i <- which(file_names %in% ignore)
-  if(length(i) > 0){
-    file_names <- file_names[-i]
-  }
-  ## Should we report the tag name too? x$tag_name?
-  local_filename(file_names)
-
+  df <- pb_info(repo, tag, .token)
+  df[c("file_name", "tag", "timestamp", "owner", "repo")]
 }
 
 #' Delete an asset attached to a release
@@ -358,19 +348,15 @@ pb_delete <- function(file,
                       repo = guess_repo(),
                       tag = "latest",
                       .token = get_token()){
-  x <- release_info(repo, tag, .token)
-
+  df <- pb_info(repo, tag, .token)
   name <- asset_filename(file)
 
-  filenames <- vapply(x$assets, `[[`, character(1), "name")
-  ids <- vapply(x$assets, `[[`, integer(1), "id")
-
   out <- NULL
-  if(name %in% filenames){
-    i <- which(filenames == name)
+  if(name %in% df$file_name){
+    i <- which(df$file_name == name)
     ## If we find matching id, Delete file from release.
     gh::gh("DELETE /repos/:owner/:repo/releases/assets/:id",
[…]
