### 2018-09-04 ropensci/piggyback#8 @cboettig
timestamp behavior now the default for upload / download
