### 2018-09-04 ropensci/piggyback#12 @cboettig
Remaining onboarding tasks
paraphasing @mpadge from
https://github.com/ropensci/onboarding/issues/220#issuecomment-413135933

## Error messages and tags

> There nevertheless remain a couple of unhelpful error messages prompted by passing inappropriate arguments through to package dependencies, primarily `httr`. One simple one of these is:

```{r}
pb_upload ("non-existent-file.Rds", repo = "mpadge/myrepo")
```
> This still passes non-existent file name through to `httr` and returns `file.exists(path) is not TRUE`.

done!

> The remainder may be more effectively addressed through first extracting all tags for a repository and then using those throughout to check each `tag` parameter.

> This would then allow the kind of prompt I suggested in the initial review along the lines of

```{r}
pb_upload (..., tag = "non-existent-tag")
# Tag 'non-existent-tag' does not exist in that repository. Would you like it to be created?
```

done.  Actions to a non-existant tag in interactive mode will prompt to create the release tag.

> And it would happen instantly, without recourse to any API calls.

I don't see how this can be done without an API call.  There's really no valid way I can see to cache this information and be sure that it has not since been changed.

> This would also enable error messages for the following call:

```{r}
pb_new_release(repo = "mpadge/junk", tag = "this-tag-already-exists")
```

done!

## other stuff

### pb_list

> This function still does not return tag information - did you intend to incorporate that?  I would reiterate most of the comments I previously made regarding this function.

done.

### pb_delete

> As previously stated, it would still be nice to be able to delete an entire release. At the moment this is not possible:

```{r}
pb_delete (repo = "mpadge/junk", tag = "v1")
# Error in asset_filename(file) :
#   argument "file" is missing, with no default
```
> Any comments on potentially enabling that?

done!

### pb_track

> This does indeed work as @cboettig suggests, but I note that `pb_upload` is not vectorised:

```{r}
pb_track ("data/*") # where * matches >= 2 files
pb_track () %>% pb_upload (repo = "mpadge/junk", tag = "v1")
# Error in vapply(elements, encode, character(1)) :
#   values must be length 1,
#  but FUN(X[[1]]) result is length 2
```
> It would be very useful to simply make current `pb_upload` some kind of `pb_upload_one` function and then have `pb_upload` be a simple `vapply` call to that.

done!

> I'll leave my comments there for the moment, and await feedback from @cboettig

These are all super helpful comments, and more such feedback always welcome.
