### 2018-09-04 1dd8099
Add @mpadge as ctb
- allow delete to be vectorized and delete all assets if no file name is given.

- add appropriate warning about tag not found, support creating tagged release on spot if interactive()
