+++ man/piggyback-package.Rd
@@ -38,4 +38,9 @@ Useful links:
 \author{
 \strong{Maintainer}: Carl Boettiger \email{cboettig@gmail.com} (0000-0002-1642-628X) [copyright holder]
 
+Other contributors:
+\itemize{
+  \item Mark Padgham (0000-0003-2172-5265) [contributor]
+}
+
 }
