### 2018-09-06 #220 @coatless
@cboettig a few. I need to sit down and type them up. Unfortunately, I've been slammed with the semester start and meetings. I'm going to try to carve out time on Thursday/Friday.
