### 2018-09-12 501f216
Style fixes from running styler::style_pkg()
