+++ R/push-pull.R
@@ -31,14 +31,15 @@
 #' pb_track(c("*.tsv", "*.tsv.gz"))
 #'
 #' }
-pb_track <- function(glob = NULL, repo_root = usethis::proj_get()){
-
-  if(!is.null(glob)){
-    write_union(usethis::proj_get(),
-                ".pbattributes",
-                glob)
+pb_track <- function(glob = NULL, repo_root = usethis::proj_get()) {
+  if (!is.null(glob)) {
+    write_union(
+      usethis::proj_get(),
+      ".pbattributes",
+      glob
+    )
     usethis::use_build_ignore(c(".pbattributes", "manifest.json"))
-    if(!is.null(git2r::discover_repository("."))){
+    if (!is.null(git2r::discover_repository("."))) {
       usethis::use_git_ignore("manifest.json")
       usethis::use_git_ignore(glob)
     }
@@ -57,8 +58,7 @@ pb_track <- function(glob = NULL, repo_root = usethis::proj_get()){
 #' @importFrom fs dir_ls
 #' @importFrom jsonlite write_json
 #' @importFrom magrittr %>%
-create_manifest <- function(manifest = "manifest.json"){
-
+create_manifest <- function(manifest = "manifest.json") {
   proj_dir <- usethis::proj_get()
   files <- pb_track()
 
@@ -67,22 +67,24 @@ create_manifest <- function(manifest = "manifest.json"){
   names(hashes) <- files
 
   json <- jsonlite::write_json(hashes,
-                               file.path(proj_dir, manifest),
-                               auto_unbox = TRUE,
-                               pretty = TRUE)
+    file.path(proj_dir, manifest),
+    auto_unbox = TRUE,
+    pretty = TRUE
+  )
   invisible(hashes)
 }
 
 ## Helper function
 #' @importFrom fs path_rel path_filter
-match_globs <- function(globs, proj_dir = usethis::proj_get()){
+match_globs <- function(globs, proj_dir = usethis::proj_get()) {
   unique(
-  unname(unlist(lapply(globs, function(g){
-  ## only match files (i.e. things we can hash)
-  fs::dir_ls(path = proj_dir, recursive = TRUE, type = "file") %>%
-    fs::path_rel(proj_dir) %>%
-    fs::path_filter(glob = g)
-  }))))
+    unname(unlist(lapply(globs, function(g) {
+      ## only match files (i.e. things we can hash)
+      fs::dir_ls(path = proj_dir, recursive = TRUE, type = "file") %>%
+        fs::path_rel(proj_dir) %>%
+        fs::path_filter(glob = g)
+    })))
+  )
 }
 
 #' Pull data from GitHub
@@ -135,33 +137,36 @@ pb_pull <- function(repo = guess_repo(),
                     overwrite = TRUE,
                     manifest = "manifest.json",
                     use_timestamps = FALSE,
-                    show_progress = TRUE)
-                    {
+                    show_progress = TRUE) {
 
   # Get hashes of all tracked files
   create_manifest(manifest)
   proj_dir <- usethis::proj_get()
 
   ## List files that will be newly pulled
-  files <- new_data(mode = "pull",
-                    tag = tag,
-                    manifest = manifest,
-                    repo = repo)
-  if(is.null(files)){
+  files <- new_data(
+    mode = "pull",
+    tag = tag,
+    manifest = manifest,
+    repo = repo
+  )
+  if (is.null(files)) {
     message("Already up to date")
     return(invisible(TRUE))
   }
-  if(length(files)==0){
+  if (length(files) == 0) {
     message("Already up to date")
     return(invisible(TRUE))
   }
-  pb_download(repo = repo,
-              tag = tag,
-              file = files,
-              dest = proj_dir,
-              overwrite = overwrite,
-              use_timestamps = use_timestamps,
[…]
