+++ R/gh.R
@@ -44,36 +44,38 @@ pb_download <- function(file = NULL,
                         ignore = "manifest.json",
                         use_timestamps = TRUE,
                         show_progress = TRUE,
-                        .token = get_token()){
-
+                        .token = get_token()) {
   progress <- httr::progress("down")
-  if(!show_progress){
+  if (!show_progress) {
     progress <- NULL
   }
 
 
   df <- pb_info(repo, tag, .token)
 
-  if(!is.null(file)){
+  if (!is.null(file)) {
     i <- which(df$file_name %in% file)
-    if(length(i) < 1)
-      warning(paste("file(s)",
-                 paste(crayon::blue(file), collapse=" "),
-                       "not found in repo",
-                       crayon::blue(repo)))
+    if (length(i) < 1) {
+      warning(paste(
+        "file(s)",
+        paste(crayon::blue(file), collapse = " "),
+        "not found in repo",
+        crayon::blue(repo)
+      ))
+    }
 
-    df <- df[i,]
+    df <- df[i, ]
   } else {
     i <- which(df$file_name %in% ignore)
-    if(length(i) >= 1){
-      df <- df[-i,]
+    if (length(i) >= 1) {
+      df <- df[-i, ]
     }
     file <- df$file_name
   }
 
 
   ## if dest not provided, we will write
-  if(length(dest) == 1){
+  if (length(dest) == 1) {
     i <- which(df$file_name %in% file)
     ## Make sure dest dir exists!
     dest <- fs::path_rel(file.path(dest, df$file_name[i]))
@@ -83,28 +85,26 @@ pb_download <- function(file = NULL,
   df$dest <- dest
 
 
-  if(use_timestamps){
+  if (use_timestamps) {
     local_timestamp <- fs::file_info(dest)$modification_time
     update <- df$timestamp > local_timestamp
     update[is.na(update)] <- TRUE # we'll download if missing locally
-    df <- df[update,]
+    df <- df[update, ]
 
-    if(dim(df)[[1]] < 1){
+    if (dim(df)[[1]] < 1) {
       message(paste("All files up-to-date already\n"))
     }
-
   }
 
   resp <- lapply(seq_along(df$id), function(i)
     gh_download_asset(df$owner[[1]],
-                      df$repo[[1]],
-                      id = df$id[i],
-                      destfile = dest[i],
-                      overwrite = overwrite,
-                      progress = progress)
-
-  )
- invisible(resp)
+      df$repo[[1]],
+      id = df$id[i],
+      destfile = dest[i],
+      overwrite = overwrite,
+      progress = progress
+    ))
+  invisible(resp)
 }
 
 
@@ -126,19 +126,15 @@ pb_download <- function(file = NULL,
 pb_download_url <- function(file = NULL,
                             repo = guess_repo(),
                             tag = "latest",
-                            .token = get_token()){
-
-
+                            .token = get_token()) {
   df <- pb_info(repo, tag, .token)
-  if(is.null(file)){
+  if (is.null(file)) {
     return(df$browser_download_url)
-  }else if(file %in% df$file_name){
+  } else if (file %in% df$file_name) {
     return(df[file == df$file_name, "browser_download_url"])
   } else {
     stop(paste("file", file, "not found in release", tag, "for repo", repo))
   }
-
-
 }
 
 
@@ -148,22 +144,27 @@ gh_download_asset <- function(owner,
                               repo,
                               id,
                               destfile,
-                              overwrite=TRUE,
+                              overwrite = TRUE,
                               .token = get_token(),
-                              progress = httr::progress("down")
-                              ){
-  if(fs::file_exists(destfile) && !overwrite){
[…]
