### 2018-09-12 #220 @coatless
In response to the latest points raised...

----

### Caching

Regarding caches, we must mention them in the frame of:

> There are only two hard things in Computer Science: **cache invalidation** and naming things.
>
> -- Phil Karlton

Having said this, I'm not sure caching needs to be attempted given the current rate limits. If the user is authenticated, which is a key requirement of many of `piggyback`'s functions, the number of times the API can be queried is **5000**. For users who do not supply a `GH_TOKEN`, they have **60** queries per hour (c.f. [Rate Limiting Overview](https://developer.github.com/v3/rate_limit/), [Rate Limit Rules](https://developer.github.com/v3/#rate-limiting)).

If you feel you _must_ do some level of caching, the two heavy users of `pb_info()` are `gh_download_asset()` used in `pb_download()` and `pb_upload_file()` called by `pb_upload()`. The easiest way to cache results would be with [`memoise`](https://github.com/r-lib/memoise) as @mpadge initially indicated. I say this because an option does _exist_ to supply a timeout (c.f. [`timeout()`](https://github.com/r-lib/memoise/blob/06d16ec150a6be3bcb4322e08056ee131b013b37/R/memoise.R#L164-L181)).

As an example consider:

```r
# Create a memoized pb_info function that expires after 10 seconds
# Note: The presently exposed pb_info is renamed to local_pb_info in this case
pb_info <- memoise(local_pb_info, ~timeout(10))

# Grab repo information
pb_info("cboettig/piggyback")
```

Though, to avoid another package dependency, the cache setup could be done by maintaining your own closure with a timeout procedure. Slight code sketch:

```r
# Make shift caching
repo_cache = function(timeout = getOption("pb.cache_timeout", 120)) {

  # Create internal tracking variables
  start_time <- repo_data <- repo <- tag <- .token = NULL

  # Initialize Clock
  start = function() { start_time <<- Sys.time() }

  # Split Time
  time_point <- function() {
    # Difference between NOW and start in seconds
    difftime(Sys.time(), start_time, units = "secs")
  }

  # Store the repo data
  cache_data <- function() {
    repo_data <<- piggyback:::pb_info(repo = repo, tag = tag, .token = .token)
    start()

    repo_data
  }

  # Create initial cache (public-facing)
  set_cache_info <- function(repo = guess_repo(), tag = NULL, .token = piggyback:::get_token()) {
    repo <<- repo; tag <<- tag; .token <<- .token

    cache_data()
  }

  # Compare to time stamp
  # If bad, then flush cache
  check_cache_state <- function() {
    if( time_point() > timeout ){
       cache_data()
    }
  }

  # Retrieve value and reset if need be
  get_cache <- function() {
    check_cache_state()
    repo_data
  }

  # Return functions
  list(set_cache_info = set_cache_info, get_cache = get_cache )
}

# Allocate the closure
pb_cache = repo_cache()

# Setup repo caching
pb_cache$set_cache_info("cboettig/piggyback")
pb_cache$get_cache()
```

_Misc note_: Glancing over `memoise`, it provides an impressive cloud backend to storing information on S3 + GCP that could potentially be harnessed; however, I feel that is out of the initial scope for `piggyback`.

-------

### Narrowing the API

I'm largely in agreement with sunsetting both `pb_push()` and `pb_pull()`. These functions heavily rely on an underlying mental model of the git. At that point, I think that the user would likely just commit into git already since the barrier of entry is significantly higher due to the need of a `manifest.json` file.

-------

### Package style

Style-wise, the `styler::style_pkg()` found a few issues. I've spun up a PR for you to include the changes in cboettig/piggyback#14

### Examples

Part of the `piggyback` repository is serving as a "checkpoint" for data (c.f. [`data` tag](https://github.com/cboettig/piggyback/releases/tag/data)). I wonder if such an example would be better suited inside of a secondary repository. The secondary repository could then be linked from a code example in the vignette and/or `README.md` as an example of what is possible.

------

I have a few more remarks, but the night is no longer young. I'll aim to finish round 2 later in the week.
