### 2018-09-12 ropensci/piggyback#12 @mpadge
Fabulous! This all makes the package much easier to use - thanks!
