### 2018-09-18 fec38e4
memoise working now.
timeout cannot be zero.
