+++ R/piggyback.R
@@ -15,8 +15,7 @@
 #'
 #' - [pb_upload()] / [pb_download()]:  Upload and download individual files to/from
 #'   the desired release of the specified repository
-#' - [pb_track()] + [pb_pull()] / [pb_push()]: Use a `git-lfs` style tracking of specific
-#'    file types and sync all tracked files with `pull` and `push` commands.  Convenient
-#'    for projects working with many piggybacked data files.
+#' - [pb_track()]: Use a `git-lfs` style tracking of specific
+#'    file types
 #'
 "_PACKAGE"
