+++ README.Rmd
@@ -45,7 +45,7 @@ Piggyback can download data attached to a release on any repository:
 
 ```{r results="hide"}
 library(piggyback)
-pb_download( "data/mtcars.tsv.gz", repo = "cboettig/piggyback")
+pb_download("mtcars.tsv.gz", repo = "cboettig/piggyback")
 ```
 
 
@@ -87,3 +87,6 @@ pb_download()
 Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md).
 By participating in this project you agree to abide by its terms.
 
+```{r include=FALSE}
+unlink("*.gz")
+```
