### 2018-09-18 #220 @cboettig
Okay, caching turned out to be a bit more of an adventure than I expected with memoize.

A 1 second cache on `pb_info` breaks a good number of tests and vignettes.  Many of these issues were fixed as soon as I made sure functions that update the GitHub release (`pb_upload`, `pb_release`) would also break the cache automatically.

I also wanted a way to turn off caching, turns out `memoise::timeout(0)` doesn't do what you expect, so I hacked around that.  Now cache duration can be set as an env var, `piggyback_cache_duration`, and setting it to `0` will turn caching off.

I've also sunset `pb_push` and `pb_pull`.  I've updated the docs to reflect this, also making the readme briefer and creating two separate vignettes; one with the longer intro to functions that was largely in the readme, and a separate vignette on alternative approaches to piggyback.

#### Remaining questions

One possibly significant issue I worry about for CRAN onboarding is the "write to temporary files" complaint. This is particularly likely to bite on `pb_download()` and I'm not sure how best to address it without losing the convenient default behavior based on the git repo root.  Maybe since it defaults to `usethis::proj_get()` and not `"."` they will be okay with it or won't notice?  (Arguably by setting an RStudio Project the user has declared a location(?).    `proj_get()` will throw an error if the user is not in a project, so my argument is that this is not an actual default location (unlike `here()`, which doesn't error).    Thoughts on this?
