+++ R/pb_upload.R
@@ -122,5 +122,8 @@ pb_upload_file <- function(file,
 
   cat("\n")
   httr::stop_for_status(r)
+
+  ## Release info changed, so break cache
+  memoise::forget(memoised_pb_info)
   invisible(r)
 }
