+++ tests/testthat/test-with-auth.R
@@ -56,6 +56,9 @@ test_that("working from git repo", {
 
   sink() # avoid verbose messages in test log. usethis msgs cannot turn off(?)
 
+  ## Test guessing repo
+  info <- pb_list()
+  expect_is(info, "data.frame")
   # pb_push
   library(magrittr)
   pb_track() %>%
@@ -76,6 +79,8 @@ test_that("working from git repo", {
   ## tare down
   setwd(cur)
   fs::dir_delete(tmp)
+
+
 })
 
 
@@ -102,3 +107,44 @@ test_that(
     )
   }
 )
+
+
+
+
+
+
+testthat::test_that(
+  "test delete", {
+
+    testthat::skip_if(piggyback:::get_token() == "")
+    testthat::skip_if(piggyback:::get_token() == "b2b7441daeeb010b1df26f1f60a7f1edc485e443")
+    testthat::skip_if_not(as.logical(Sys.getenv("CBOETTIG_TOKEN", FALSE)))
+
+
+    tmp <- tempdir()
+
+    readr::write_tsv(datasets::mtcars, fs::path(tmp, "mtcars2.tsv.gz"))
+
+    testthat::expect_silent(
+      out <- pb_upload(
+        file = fs::path(tmp, "mtcars2.tsv.gz"),
+        repo = "cboettig/piggyback",
+        tag = "v0.0.1",
+        overwrite = FALSE,
+        show_progress = FALSE,
+        use_timestamps = TRUE,
+        dir = tmp
+      )
+    )
+
+    pb_delete(
+      repo = "cboettig/piggyback",
+      file = "mtcars2.tsv.gz",
+      tag = "v0.0.1"
+    )
+
+    unlink(fs::path(tmp, "mtcars2.tsv.gz"))
+  }
+)
+
+
