+++ vignettes/intro.Rmd
@@ -155,7 +155,7 @@ pb_download()
 
 By default, files will not be transferred if the timestamp is more recent at the destination (e.g. `pb_download()` will not download a copy if the local version is newer than the uploaded version).  If necessary, you can override this behavior by setting `use_timestamps = FALSE` in `pb_download()` or `pb_upload()`.  
 
-To reduce API calls to GitHub, piggyback caches most calls with a timeout of 1 second by default.  This avoids repeating identical requests to update it's internal record of the repository data (releases, assets, timestamps, etc) during programmatic use.  You can increase or decrease this delay by setting the environmental variable in seconds, e.g. `Sys.setenv("piggyback_cache_duration"=10)` for a longer delay.  Note that the delay cannot be less than `1e-6` seconds at this time.
+To reduce API calls to GitHub, piggyback caches most calls with a timeout of 1 second by default.  This avoids repeating identical requests to update it's internal record of the repository data (releases, assets, timestamps, etc) during programmatic use.  You can increase or decrease this delay by setting the environmental variable in seconds, e.g. `Sys.setenv("piggyback_cache_duration"=10)` for a longer delay or `Sys.setenv("piggyback_cache_duration"=0)` to disable caching. 
 
 
 ## Path names
