### 2018-09-19 1d8de8a
more robust pb_info & pb_new_release
- handle case of no releases
- handle case of releases with no assets
