### 2018-09-19 7cd382b
migrating tests to target piggyback-testing
Uses a clean testing repo instead of cluttering releases of the actual package (WIP)
