+++ inst/examples/upload-test-files.R
@@ -9,3 +9,8 @@ pb_list(repo)
 
 write_tsv(iris, "iris2.tsv.gz")
 pb_upload( "iris2.tsv.gz", repo = repo, tag = "v0.0.1")
+
+dir.create("data", showWarnings = FALSE)
+write_tsv(mtcars, "data/mtcars.tsv.gz")
+
+pb_upload( "data/mtcars.tsv.gz", repo = repo, tag = "v0.0.1")
