### 2018-09-19 #220 @mpadge
Thanks @cboettig for doing the hard work for us all on `memoise::timeout()` - very useful to know! And as you know, I think the retirement of `pb_push/pull` is a strong boost for ease and consistency of use - thanks.

As for file writing, I don't have any great insights, and would love to be assured that what you do were permissible. I note only that:
1. I just scoured the code for `usethis`, and it seems they exclusively use `file_temp()` for writing **with default `tmp_dir = tempdir()` at all times** (yes, even when testing `proj_get()`, as in [this example](https://github.com/r-lib/usethis/blob/master/tests/testthat/test-proj.R#L71-L92)); and
2. CRAN have very strong auto-checks for any attempts to write to non-standard locations which have caught me out previously, so current usage will definitely be noticed. Have you tried testing on all [`rhub` platforms](https://github.com/r-hub/rhub#listing-r-hub-platforms) to see whether any of those flag anything?
