### 2018-09-19 #220 @cboettig
@mpadge Thanks for your thoughts on this.  Yeah, `pb_download` already supports specifying `dest = tempdir()` while defaulting to `dest = usethis::proj_get()`.

I've tried to consistently make sure tests, examples, and vignettes use tempdir in the calls (well, at least tests, I probably need another pass over this to add it to examples and vignettes).  I don't really like this because it makes the examples look more cumbersome than they really are, but that is somewhat unavoidable I guess.  (It's not just `tempdir()`; following @coatless 's very excellent point that I should use a different repo for testing, but this means that all functions include a manual declaration of `repo = "cboettig/piggyback-tests"` as well now.

My question is though, do you think that's good enough?  I'm worried that having any default location at all is technically a violation of policy.  (e.g. in base function `download.file()`, `dest` has no default).

(Though I still feel this is not technically a default location, since `proj_get()` throws an error unless the user has set a project).
