+++ README.Rmd
@@ -45,13 +45,13 @@ Piggyback can download data attached to a release on any repository:
 
 ```{r results="hide"}
 library(piggyback)
-pb_download("data/mtcars.tsv.gz", repo = "cboettig/piggyback-tests")
+pb_download("data/mtcars.tsv.gz", repo = "cboettig/piggyback-tests", dest = tempdir())
 ```
 
 
-Downloading from private repos or uploading to any repo requires authentication, so be sure to set a `GITHUB_TOKEN` (or `GITHUB_PAT`) environmental variable, or include the `.token` argument. See [vignette](https://cboettig.github.io/piggyback/articles/intro.html) or function documentation for details.
+Downloading from private repos or uploading to any repo requires authentication, so be sure to set a `GITHUB_TOKEN` (or `GITHUB_PAT`) environmental variable, or include the `.token` argument.  Omit the file name to download all attached objects. Omit the repository name to default to the current repository.  See [vignette](https://cboettig.github.io/piggyback/articles/intro.html) or function documentation for details.  
 
-We can then also upload data to an existing release:
+We can also upload data to any existing release (defaults to `latest`):
 
 ```{r eval=FALSE}
 ## We'll need some example data first.
@@ -67,7 +67,6 @@ pb_upload("mtcars.tsv.gz", repo = "cboettig/piggyback-tests")
 For a [Git LFS](https://git-lfs.github.com/) style workflow, just specify the type of files you wish to track using `pb_track()`.  Piggyback will retain a record of these files in a hidden `.pbattributes` file in your repository, and add these to `.gitignore` so you don't accidentally commit them to GitHub.  `pb_trak` will aslo return a list of such files that you can easily pass to `pb_upload()`:  
 
 ```{r eval=FALSE}
-library(piggyback)
 # track csv files, compressed data, and geotiff files:
 pb_track(c("*.csv", "*.gz", "*.tif")) %>%
 pb_upload()
