+++ vignettes/intro.Rmd
@@ -17,7 +17,7 @@ knitr::opts_chunk$set(
   eval = Sys.getenv("CBOETTIG_TOKEN", FALSE)
 )
 
-Sys.setenv(piggyback_cache_duration="1e-6")
+Sys.setenv(piggyback_cache_duration=0)
 
 ```
 
@@ -55,22 +55,26 @@ Download the latest version or a specific version of the data:
 
 ```{r}
 library(piggyback)
-pb_download("iris2.tsv.gz", repo = "cboettig/piggyback-tests")
+pb_download("iris2.tsv.gz", 
+            repo = "cboettig/piggyback-tests",
+            tag = "v0.0.1",
+            dest = tempdir())
 ```
 
-**Note**: Whenever you are working from a location inside a git repository corresponding to your GitHub repo, you can simply omit the `repo` argument and it will be detected automatically.  
+**Note**: Whenever you are working from a location inside a git repository corresponding to your GitHub repo, you can simply omit the `repo` argument and it will be detected automatically. Likewise, if you omit the release `tag`, `the`pb_download` will simply pull data from most recent release (`latest`).  Third, you can omit `tempdir()` if you are using an RStudio Project (`.Rproj` file) in your repository, and then the download location will be relative to Project root.  `tempdir()` is used throughout the examples only to meet CRAN policies and is unlikely to be the choice you actually want here.  
 
-Or a specific version:
 
-```{r}
-pb_download("iris2.tsv.gz", repo = "cboettig/piggyback-tests", tag = "v0.0.1")
-```
-
-Or simply omit the file name to download all assets connected with a given release.  (You can also always specify a destination directory to download. This directory will be created if it does not yet exist.)
+Lastly, simply omit the file name to download all assets connected with a given release.  
 
 ```{r results="hide"}
-pb_download(repo = "cboettig/piggyback-tests")
+pb_download(repo = "cboettig/piggyback-tests",
+            tag = "v0.0.1",
+            dest = tempdir())
 ```  
+
+These defaults mean that in most cases, it is sufficient to simply call `pb_download()` without additional arguments to pull in any data associated with a project on a GitHub repo that is too large to commit to git directly. 
+
+`pb_download()` will skip the download of any file that already exists locally if the timestamp on the local copy is more recent than the timestamp on the GitHub copy.  `pb_download()` also includes arguments to control the timestamp behavior, progress bar, whether existing files should be overwritten, or if any particular files should not be downloaded.  See function documentation for details.  
 
 
 Sometimes it is preferable to have a URL from which the data can be read in directly, rather than downloading the data to a local file.  For example, such a URL can be embedded directly into another R script, avoiding any dependence on `piggyback` (provided the repository is already public.)  To get a list of URLs rather than actually downloading the files, use `pb_download_url()`:
@@ -81,14 +85,6 @@ pb_download_url("data/mtcars.tsv.gz",
                 tag = "v0.0.1") 
 ```
 
-Or store and read the URL directly into R:
-
-```{r eval=FALSE}
-url <- pb_download_url("data/mtcars.tsv.gz", 
-                       repo = "cboettig/piggyback-tests",
-                       tag = "v0.0.1")
-readr::read_tsv(url)
-```
 
 
 ## Uploading data
@@ -106,34 +102,37 @@ Once we have at least one release available, we are ready to upload.  By default
 ## Pro tip: compress your tabular data to save space & speed upload/downloads
 readr::write_tsv(mtcars, "mtcars.tsv.gz")
 
-pb_upload("mtcars.tsv.gz", repo = "cboettig/piggyback-tests", tag = "v0.0.1")
+pb_upload("mtcars.tsv.gz", 
+          repo = "cboettig/piggyback-tests", 
+          tag = "v0.0.1")
 ```
 
[…]
