### 2018-09-20 #220 @cboettig
@annakrystalli @mpadge @coatless

Okay, I think I basically have things where I want them now in response to all the super excellent feedback.  You have all made this a way more usable, useful, and robust package.  I'm sure it's not perfect, but I think we've hammered out a nice API and functionality to get people started.

I think we have the :+1: from Mark to onboard, but still waiting final confirmation from James, yes?

@annakrystalli Remind me any additional steps I need to do for the JOSS part? I just go over there and open an issue pointing back here?  Mark/Jim, are you :+1: with the JOSS submission?
