+++ NAMESPACE
@@ -27,6 +27,7 @@ importFrom(httr,stop_for_status)
 importFrom(httr,upload_file)
 importFrom(httr,write_disk)
 importFrom(jsonlite,toJSON)
+importFrom(lubridate,as_datetime)
 importFrom(magrittr,"%>%")
 importFrom(memoise,forget)
 importFrom(memoise,memoise)
