### 2018-09-21 #220 @cboettig
@stefaniebutland Thanks, yes I'd love to do a blog post on `piggyback`, maybe combine a bit of `arkdb`.  I was thinking I might focus on why I wrote these two little packages, more than the technical details (each package basically has only two functions, essentially data_in and data_out), so they are conceptually simple anyway.
