+++ DESCRIPTION
@@ -18,8 +18,8 @@ Authors@R: c(person("Carl", "Boettiger",
              person("Mark", "Padgham", 
                      role = c("ctb"),
                      comment = c(ORCID = "0000-0003-2172-5265")))
-URL: https://github.com/cboettig/piggyback
-BugReports: https://github.com/cboettig/piggyback/issues
+URL: https://github.com/ropensci/piggyback
+BugReports: https://github.com/ropensci/piggyback/issues
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
