+++ README.md
@@ -34,13 +34,13 @@ You can install the development version from
 
 ``` r
 # install.packages("devtools")
-devtools::install_github("cboettig/piggyback")
+devtools::install_github("ropensci/piggyback")
 ```
 
 ## Quickstart
 
 See the [piggyback
-vignette](https://cboettig.github.io/piggyback/articles/intro.html) for
+vignette](https://ropensci.github.io/piggyback/articles/intro.html) for
 details on authentication and additional package functionality.
 
 Piggyback can download data attached to a release on any repository:
@@ -55,7 +55,7 @@ authentication, so be sure to set a `GITHUB_TOKEN` (or `GITHUB_PAT`)
 environmental variable, or include the `.token` argument. Omit the file
 name to download all attached objects. Omit the repository name to
 default to the current repository. See
-[vignette](https://cboettig.github.io/piggyback/articles/intro.html) or
+[vignette](https://ropensci.github.io/piggyback/articles/intro.html) or
 function documentation for details.
 
 We can also upload data to any existing release (defaults to `latest`):
