+++ vignettes/intro.Rmd
@@ -33,7 +33,7 @@ You can install the development version from [GitHub](https://github.com/) with:
 
 ``` r
 # install.packages("devtools")
-devtools::install_github("cboettig/piggyback")
+devtools::install_github("ropensci/piggyback")
 ```
 
 ## Authentication
