+++ README.Rmd
@@ -10,6 +10,8 @@ output: github_document
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/cboettig/piggyback?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/piggyback)
 [![CRAN status](https://www.r-pkg.org/badges/version/piggyback)](https://cran.r-project.org/package=piggyback)
 [![](https://badges.ropensci.org/220_status.svg)](https://github.com/ropensci/onboarding/issues/220)
+[![DOI](https://zenodo.org/badge/132979724.svg)](https://zenodo.org/badge/latestdoi/132979724)
+
 
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
