+++ NEWS.md
@@ -1,3 +1,12 @@
+
+# piggyback 0.0.6 
+
+* bugfix for migrating unit test
+
+# piggyback 0.0.5 2018-09-21
+
+* initial Onboarding to rOpenSci
+
 # piggyback 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
