### 2018-09-22 02b261e
examples should also only test against test repo!
