+++ man/pb_download_url.Rd
@@ -32,8 +32,8 @@ dependency on \code{piggyback} or authentication steps.
 \examples{
 \donttest{
 
-pb_download_url("data/iris.tsv.gz",
-                repo = "cboettig/piggyback",
+pb_download_url("data/iris.tsv.xz",
+                repo = "cboettig/piggyback-tests",
                 tag = "v0.0.1")
 
 }
