+++ man/pb_download.Rd
@@ -47,9 +47,11 @@ Download data from an existing release
 \examples{
 \donttest{
  ## Download a specific file:
- piggyback::pb_download("data/iris.tsv.gz", repo = "cboettig/piggyback")
+ piggyback::pb_download("data/iris.tsv.gz",
+                        repo = "cboettig/piggyback-tests")
 
  ## Download all files
- piggyback::pb_download(repo = "cboettig/piggyback")
+ piggyback::pb_download(repo = "cboettig/piggyback-tests")
+
 }
 }
