+++ R/pb_download.R
@@ -27,12 +27,16 @@
 #' @importFrom fs dir_create
 #' @export
 #' @examples \donttest{
-#'  ## Download a specific file:
+#'  ## Download a specific file.
+#'  ## (dest can be omitted when run inside and R project)
 #'  piggyback::pb_download("data/iris.tsv.gz",
-#'                         repo = "cboettig/piggyback-tests")
-#'
+#'                         repo = "cboettig/piggyback-tests",
+#'                         dest = tempdir())
+#' }
+#' \dontrun{
 #'  ## Download all files
-#'  piggyback::pb_download(repo = "cboettig/piggyback-tests")
+#'  piggyback::pb_download(repo = "cboettig/piggyback-tests",
+#'                         dest = tempdir())
 #'
 #' }
 pb_download <- function(file = NULL,
