### 2018-09-22 290d3ab
winbuilder refuses to obey knitr eval commands
