### 2018-09-24 94fbdde
Update paper.md following @arfon's suggestions
