+++ paper.bib
@@ -1,7 +1,7 @@
 @MISC{GitLFS,
 url = {https://git-lfs.github.com/},
 canonicaluri = {\url{https://git-lfs.github.com/}},
-title = {Git LFS},
+title = {{Git LFS}},
 YEAR = {2018},
 howpublished = {\url{https://git-lfs.github.com/}},
 authors = {{GitHub Inc.}}
@@ -11,9 +11,43 @@ authors = {{GitHub Inc.}}
 @MISC{API3,
 url = {https://developer.github.com/v3/},
 canonicaluri = {\url{https://developer.github.com/v3/}},
-title = {GitHub API version 3},
+title = {{GitHub API version 3}},
 YEAR = {2018},
 howpublished = {\url{https://developer.github.com/v3/}},
 authors = {{GitHub Inc.}}
+}
+
+@Article{Ram2013,
+author="Ram, Karthik",
+title="Git can facilitate greater reproducibility and increased transparency in science",
+journal="Source Code for Biology and Medicine",
+year="2013",
+month="Feb",
+day="28",
+volume="8",
+number="1",
+pages="7",
+issn="1751-0473",
+doi="10.1186/1751-0473-8-7",
+url="https://doi.org/10.1186/1751-0473-8-7"
+}
+
+@MISC{Boettiger2018,
+author = "Carl Boettiger",
+url = {https://ropensci.github.io/piggyback/articles/alternatives.html},
+canonicaluri = {\url{https://ropensci.github.io/piggyback/articles/alternatives.html}},
+title = {Piggyback comparison to alternatives},
+year = {2018},
+archived = {http://greycite.knowledgeblog.org/?uri=https%3A%2F%2Fropensci.github.io%2Fpiggyback%2Farticles%2Falternatives.html},
+}
+
+
+@MISC{Boettiger2018b,
+author = "Carl Boettiger",
+url = {https://ropensci.github.io/piggyback/articles/intro.html},
+canonicaluri = {\url{https://ropensci.github.io/piggyback/articles/intro.html}},
+title = {{Piggyback Data atop your GitHub Repository!}},
+year = {2018},
+archived = {http://greycite.knowledgeblog.org/?uri=https%3A%2F%2Fropensci.github.io%2Fpiggyback%2Farticles%2Fintro.html},
 }
 
