### 2018-09-24 2a88ccf
Have pb_list include file size info
